Refactored the HtmlAgilityPack operations out of the writer project and into common. Added integration test to common tests. Added test resource
