+++ STSCommon.Test/Utilities/StringUtilitiesTests.cs
@@ -8,45 +8,6 @@ namespace STSCommon.Test.Utilities
     public class StringUtilitiesTests
     {
         [Test]
-        public void MatchCharactersInRangeValidMatchShouldReturnTrue()
-        {
-            // Arrange
-            const string input = "A";
-
-            // Act
-            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
-
-            // Assert
-            Assert.IsTrue(result);
-        }
-
-        [Test]
-        public void MatchCharactersInRangeValidWithExtraWhitespaceMatchShouldReturnTrue()
-        {
-            // Arrange
-            const string input = "  A  \n\t";
-
-            // Act
-            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
-
-            // Assert
-            Assert.IsTrue(result);
-        }
-
-        [Test]
-        public void MatchCharactersInRangeUnmatchedCharShouldReturnFalse()
-        {
-            // Arrange
-            const string input = "a";
-
-            // Act
-            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
-
-            // Assert
-            Assert.IsFalse(result);
-        }
-
-        [Test]
         public void MatchCharactersInRangeLongStringShouldReturnFalse()
         {
             // Arrange
@@ -67,6 +28,45 @@ namespace STSCommon.Test.Utilities
 
             // Act
             var result = input.All(x => StringUtilities.MatchesCharacterInRange(x.ToString(), 'A', 'K'));
+
+            // Assert
+            Assert.IsTrue(result);
+        }
+
+        [Test]
+        public void MatchCharactersInRangeUnmatchedCharShouldReturnFalse()
+        {
+            // Arrange
+            const string input = "a";
+
+            // Act
+            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
+
+            // Assert
+            Assert.IsFalse(result);
+        }
+
+        [Test]
+        public void MatchCharactersInRangeValidMatchShouldReturnTrue()
+        {
+            // Arrange
+            const string input = "A";
+
+            // Act
+            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
+
+            // Assert
+            Assert.IsTrue(result);
+        }
+
+        [Test]
+        public void MatchCharactersInRangeValidWithExtraWhitespaceMatchShouldReturnTrue()
+        {
+            // Arrange
+            const string input = "  A  \n\t";
+
+            // Act
+            var result = StringUtilities.MatchesCharacterInRange(input, 'A', 'A');
 
             // Assert
             Assert.IsTrue(result);
