+++ STSParser.Test/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
+
 <packages>
   <package id="HtmlAgilityPack" version="1.4.9.5" targetFramework="net462" />
   <package id="NUnit" version="3.6.1" targetFramework="net462" />
