--- STSWriter/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<packages>
-  <package id="HtmlAgilityPack" version="1.4.9.5" targetFramework="net462" />
-</packages>
